feature: Implement calculate for indirect flow

+++ sep490-enterprise/src/main/java/greenbuildings/enterprise/dtos/EmissionActivityRecordDTO.java
@@ -15,6 +15,6 @@ public record EmissionActivityRecordDTO(
         LocalDate endDate,
         RecordFileDTO file,
         int quantity,
-        double ghg
+        BigDecimal ghg
 ) {
 }
